Fix #285: LTO compilation warnings
Fix new warnings displayed when compiling with LTO.

+++ pcp_stats.c
@@ -30,7 +30,7 @@
 #define _(string) (string)
 #endif
 
-extern unsigned int flags;
+extern uint64_t flags;
 
 #ifdef HAVE_PCP
 #include <pcp/pmapi.h>
