+++ xml_stats.c
@@ -35,7 +35,7 @@
 #define _(string) (string)
 #endif
 
-extern unsigned int flags;
+extern uint64_t flags;
 
 /*
  ***************************************************************************
